review fixes - rename package
