+++ src/main/java/pages/LinkUrl.java
@@ -1,4 +1,4 @@
-package pageObject;
+package pages;
 
 public class LinkUrl {
     public static final String BASE_URL = "https://stellarburgers.nomoreparties.site/";
