+++ src/main/java/pages/MainPage.java
@@ -1,4 +1,4 @@
-package pageObject;
+package pages;
 
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
