+++ src/main/java/pages/RestorePasswordPage.java
@@ -1,4 +1,4 @@
-package pageObject;
+package pages;
 
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
